Made bullet types work in iso
Had to edit iso player a bit to make the bullet type work with it. firstly made it inherit the player class. This also means a lot of the code with functions in the script could be deleted. refactored heath again so it is within the class. added an extra argument into the bullet type script so it can handle isometric transform.